requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.1.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
Django==3.0.9
django-celery-beat==2.1.0
django-cors-headers==3.4.0
django-dbbackup==3.3.0
django-dkim==0.1.2
django-filter==2.3.0
django-gdrive-backup==0.0.2
django-nested-inline==0.4.4
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dkimpy==1.0.5
dnspython==1.16.0
encrypted-credentials==0.0.2
eventlet==0.29.1
Faker==8.9.1
firebase-admin==5.0.2
future==0.18.2
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-client-helper==0.0.3
google-cloud-core==1.7.2
google-cloud-firestore==2.1.3
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.52.0
greenlet==0.4.17
grpcio==1.39.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
kombu==4.6.11
Markdown==3.2.2
msgpack==1.0.2
oauth2client==4.1.3
packaging==21.0
Pillow==7.2.0
proto-plus==1.19.0
protobuf==3.13.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydkim==0.3
PyJWT==1.7.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
